[#3] signature: Refactor sign and verify
Split methods to separate files, drop redundant intefaces
